refactor: 공통으로 사용되는 ProfileImage, PostFeed -> designsystem 모듈로 이동
