+++ core/designsystem/build.gradle.kts
@@ -9,6 +9,7 @@ android {
 
 dependencies {
     implementation(projects.core.common)
+    implementation(projects.core.domain)
 
     implementation(libs.coil.compose)
 }
