+++ feature/home/src/main/java/com/example/home/graph/post/component/CommentView.kt
@@ -11,9 +11,7 @@ import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.layout.size
 import androidx.compose.foundation.layout.width
-import androidx.compose.foundation.shape.CircleShape
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.getValue
@@ -22,19 +20,14 @@ import androidx.compose.runtime.remember
 import androidx.compose.runtime.setValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.draw.clip
 import androidx.compose.ui.graphics.ColorFilter
-import androidx.compose.ui.layout.ContentScale
-import androidx.compose.ui.platform.LocalContext
 import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
-import coil3.compose.rememberAsyncImagePainter
-import coil3.request.ImageRequest
-import coil3.request.crossfade
 import com.example.common.util.clickable
 import com.example.designsystem.R
+import com.example.designsystem.component.ProfileImage
 import com.example.designsystem.theme.Background
 import com.example.designsystem.theme.DarkGray
 import com.example.designsystem.theme.Gray
@@ -56,13 +49,6 @@ internal fun CommentView(
     var isOwnCommentDropDownMenuExpanded by remember { mutableStateOf(false) }
     var isOtherCommentDropDownMenuExpanded by remember { mutableStateOf(false) }
 
-    val profileImage = rememberAsyncImagePainter(
-        model = ImageRequest.Builder(LocalContext.current)
-            .data(comment.profileImageUrl ?: R.drawable.default_profile)
-            .crossfade(true)
-            .build()
-    )
-
     val backgroundColor =
         if (replyTargetId != null && replyTargetId.equals(comment.commentId)) PrimaryDefault.copy(
             alpha = 0.2f
@@ -76,13 +62,10 @@ internal fun CommentView(
                 modifier = Modifier.fillMaxWidth(),
                 verticalAlignment = Alignment.CenterVertically
             ) {
-                Image(
-                    painter = profileImage,
-                    contentDescription = "프로필 이미지",
-                    modifier = Modifier
-                        .size(25.dp)
-                        .clip(CircleShape),
-                    contentScale = ContentScale.Crop
+                ProfileImage(
+                    profileImageUrl = comment.profileImageUrl,
+                    imageSize = if (comment.profileImageUrl != null) 23.dp else 21.dp,
+                    paddingValue = if (comment.profileImageUrl != null) 1.dp else 2.dp
                 )
 
                 Spacer(Modifier.width(6.dp))
@@ -176,14 +159,6 @@ private fun ChildCommentView(
     onDelete: (Int) -> Unit,
     onReport: (Int) -> Unit
 ) {
-
-    val profileImage = rememberAsyncImagePainter(
-        model = ImageRequest.Builder(LocalContext.current)
-            .data(comment.profileImageUrl ?: R.drawable.default_profile)
-            .crossfade(true)
-            .build()
-    )
-
     var isOwnCommentDropDownMenuExpanded by remember { mutableStateOf(false) }
     var isOtherCommentDropDownMenuExpanded by remember { mutableStateOf(false) }
 
@@ -196,13 +171,10 @@ private fun ChildCommentView(
             modifier = Modifier.fillMaxWidth(),
             verticalAlignment = Alignment.CenterVertically
         ) {
-            Image(
-                painter = profileImage,
-                contentDescription = "프로필 이미지",
-                modifier = Modifier
-                    .size(25.dp)
-                    .clip(CircleShape),
-                contentScale = ContentScale.Crop
+            ProfileImage(
+                profileImageUrl = comment.profileImageUrl,
+                imageSize = if (comment.profileImageUrl != null) 23.dp else 19.dp,
+                paddingValue = if (comment.profileImageUrl != null) 1.dp else 3.dp
             )
 
             Spacer(Modifier.width(6.dp))
